+++ src/activity.h
@@ -302,6 +302,8 @@ public:
     double ret = std::abs(get<double>(value, 0, arrival));
     if (ret == R_PosInf) ret = -1;
     get_resource(arrival)->set_capacity((int)ret);
+    if (arrival->is_paused())
+      return ENQUEUE;
     return 0;
   }
 
@@ -733,7 +735,7 @@ public:
       if (i < heads.size())
         selected = heads[i];
       Arrival* new_arrival = arrival->clone();
-      new_arrival->set_activity(this->get_next());
+      new_arrival->set_activity(get_next());
       new_arrival->activate();
     }
     if (heads.size())
@@ -860,7 +862,7 @@ protected:
     if (!(*ptr) || *ptr != target)
       return;
     if ((*ptr)->size()) {
-      (*ptr)->set_activity(this->get_next());
+      (*ptr)->set_activity(get_next());
       (*ptr)->activate();
       *ptr = init(*ptr);
     } else {
@@ -889,7 +891,7 @@ public:
     Batched* batched = dynamic_cast<Batched*>(arrival);
     if (!batched || batched->is_permanent())
       return 0;
-    batched->pop_all(this->get_next());
+    batched->pop_all(get_next());
     delete batched;
     return REJECT;
   }
@@ -1058,7 +1060,7 @@ protected:
   UMAP<Arrival*, Activity*> pending;
 
   void launch_handler(Arrival* arrival) {
-    if (!arrival->is_active())
+    if (!arrival->sim->is_scheduled(arrival))
       return;
     arrival->stop();
     if (heads.size()) {