+++ tests/testthat/test-trajectory-set-capacity-queue.R
@@ -113,3 +113,20 @@ test_that("queue is dropped with queue_size_strict (2)", {
   expect_equal(run(env, 1) %>% get_queue_count("res"), 3)
   expect_equal(run(env, 2) %>% get_queue_count("res"), 0)
 })
+
+test_that("self-induced preemption works", {
+  t <- trajectory() %>%
+    seize("res") %>%
+    set_capacity("res", 0) %>%
+    timeout(10) %>%
+    release("res")
+
+  env <- simmer(verbose = TRUE) %>%
+    add_resource("res", preemptive = TRUE) %>%
+    add_generator("dummy", t, at(0)) %>%
+    run(5)
+
+  expect_equal(get_queue_count(env, "res"), 1)
+  expect_equal(now(env), 0)
+  expect_equal(peek(env), numeric(0))
+})